loops better

+++ kernel/train_state.h
@@ -17,14 +17,14 @@ struct TrainState {
   bool light_on : 1 = true;
 
   // units of um/tick (micrometer per tick)
-  std::array<uint16_t, 15> v_max{0,   0,   32,  40,  80,  100, 140, 200,
+  std::array<uint32_t, 15> v_max{0,   0,   32,  40,  80,  100, 140, 200,
                                  250, 310, 360, 440, 500, 532, 579};
   // units of nm/ticks^2 (nanometer per tick^2) aka 1000*um / ticks^2
   std::array<int, 15> accel{
       -47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47, 47,
   };
   // units of um / tick
-  uint16_t ve{0};
+  uint32_t ve{0};
 };
 
 struct State {
